+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/builtins/objects/type/TypeNodes.java
@@ -222,6 +222,7 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.CompilerDirectives.ValueType;
 import com.oracle.truffle.api.HostCompilerDirectives.InliningCutoff;
 import com.oracle.truffle.api.RootCallTarget;
+import com.oracle.truffle.api.TruffleLanguage.Env;
 import com.oracle.truffle.api.dsl.Bind;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Cached.Exclusive;
@@ -237,6 +238,8 @@ import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.frame.Frame;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.UnknownIdentifierException;
+import com.oracle.truffle.api.interop.UnsupportedMessageException;
 import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.api.nodes.ControlFlowException;
 import com.oracle.truffle.api.nodes.Node;
@@ -1482,23 +1485,23 @@ public abstract class TypeNodes {
         }
 
         @Specialization
-        static boolean doManaged(PythonBuiltinClassType left, PythonBuiltinClassType right) {
+        static boolean doTypeType(PythonBuiltinClassType left, PythonBuiltinClassType right) {
             return left == right;
         }
 
         @Specialization
-        static boolean doManaged(PythonBuiltinClassType left, PythonBuiltinClass right) {
+        static boolean doTypeClass(PythonBuiltinClassType left, PythonBuiltinClass right) {
             return left == right.getType();
         }
 
         @Specialization
-        static boolean doManaged(PythonBuiltinClass left, PythonBuiltinClassType right) {
+        static boolean doClassType(PythonBuiltinClass left, PythonBuiltinClassType right) {
             return left.getType() == right;
         }
 
         @Specialization
         @InliningCutoff
-        static boolean doNativeSingleContext(PythonAbstractNativeObject left, PythonAbstractNativeObject right,
+        static boolean doNative(PythonAbstractNativeObject left, PythonAbstractNativeObject right,
                         @CachedLibrary(limit = "1") InteropLibrary lib) {
             if (left == right) {
                 return true;
@@ -1507,6 +1510,44 @@ public abstract class TypeNodes {
                 return (long) left.getPtr() == (long) right.getPtr();
             }
             return lib.isIdentical(left.getPtr(), right.getPtr(), lib);
+        }
+
+        @Specialization(guards = {"isForeignObject(left)", "isForeignObject(right)"})
+        @InliningCutoff
+        static boolean doOther(Object left, Object right,
+                        @Bind PythonContext context,
+                        @CachedLibrary(limit = "2") InteropLibrary lib) {
+            if (lib.isMetaObject(left) && lib.isMetaObject(right)) {
+                if (left == right) {
+                    return true;
+                }
+                // *sigh*... Host classes have split personality with a "static" and a "class"
+                // side, and that affects identity comparisons. And they report their "class" sides
+                // as bases, but importing from Java gives you the "static" side.
+                Env env = context.getEnv();
+                if (env.isHostObject(left) && env.isHostObject(right)) {
+                    // the activation of isMemberReadable and later readMember serves as branch
+                    // profile
+                    boolean leftIsStatic = lib.isMemberReadable(left, "class");
+                    boolean rightIsStatic = lib.isMemberReadable(right, "class");
+                    if (leftIsStatic != rightIsStatic) {
+                        try {
+                            if (leftIsStatic) {
+                                left = lib.readMember(left, "class");
+                            } else {
+                                assert rightIsStatic;
+                                right = lib.readMember(right, "class");
+                            }
+                        } catch (UnsupportedMessageException | UnknownIdentifierException e) {
+                            throw CompilerDirectives.shouldNotReachHere(e);
+                        }
+                    }
+                }
+                if (lib.isIdentical(left, right, lib)) {
+                    return true;
+                }
+            }
+            return false;
         }
 
         @Fallback