+++ graalpython/lib-graalpython/__graalpython__.py
@@ -142,7 +142,17 @@ def import_current_as_named_module_with_delegate(module, module_name, delegate_n
 
 
 @builtin
-def build_java_class(module, ns, name, base):
+def build_java_class(module, ns, name, base, new_style=False):
+    if new_style:
+        return build_new_style_java_class(ns, name, base)
+    import warnings
+    warnings.warn("Subclassing Java classes is going to change "
+                  "to a new instance layout that is hopefully "
+                  "more intuitive. Pass the keyword new_style=True "
+                  "to your class definition to try the new style. "
+                  "The new style will become the default in the next "
+                  "release and the old style will be removed soon after.", DeprecationWarning, 1)
+
     ns['__super__'] = None  # place where store the original java class when instance is created
     ExtenderClass = type("PythonJavaExtenderClass", (object, ), ns)
     HostAdapter = __graalpython__.extend(base)
@@ -158,3 +168,115 @@ def build_java_class(module, ns, name, base):
 
     resultClass.__new__ = classmethod(factory)
     return resultClass
+
+
+_CUSTOM_JAVA_SUBCLASS_BACKSTOPS = {}
+
+
+@builtin
+def build_new_style_java_class(module, ns, name, base):
+    import polyglot
+    import types
+
+    JavaClass = __graalpython__.extend(base)
+    if JavaClass not in _CUSTOM_JAVA_SUBCLASS_BACKSTOPS:
+        class MroClass:
+            def __getattr__(self, name):
+                sentinel = object()
+                # An attribute access on the Java instance failed, check the
+                # delegate and then the static Java members
+                result = getattr(self.this, name, sentinel)
+                if result is sentinel:
+                    return getattr(self.getClass().static, name)
+                else:
+                    return result
+
+            def __setattr__(self, name, value):
+                # An attribute access on the Java instance failed, use the delegate
+                setattr(self.this, name, value)
+
+            def __delattr__(self, name):
+                # An attribute access on the Java instance failed, use the delegate
+                delattr(self.this, name)
+
+        # This may race, so we allow_method_overwrites, at the only danger to
+        # insert a few useless classes into the MRO
+        polyglot.register_interop_type(JavaClass, MroClass, allow_method_overwrites=True)
+
+    # A class to make sure that the returned Python class can be used for
+    # issubclass and isinstance checks with the Java instances, and to wrap all
+    # methods in created subclasses to provide the proper `self` (the Java
+    # instance) and still make `super()` work.
+    class JavaSubclassMeta(type):
+        @property
+        def __bases__(self):
+            return (JavaClass,)
+
+        def __instancecheck__(cls, obj):
+            return isinstance(obj, JavaClass)
+
+        def __subclasscheck__(cls, derived):
+            return cls is derived or issubclass(derived, JavaClass)
+
+        def __new__(mcls, name, bases, namespace):
+            if bases:
+                new_class = None
+
+                class custom_super():
+                    def __init__(self, start_type=None, object_or_type=None):
+                        assert start_type is None and object_or_type is None, "super() calls in Python class inheriting from Java must not receive arguments"
+                        f = sys._getframe(1)
+                        self.self = f.f_locals[f.f_code.co_varnames[0]]
+
+                    def __getattribute__(self, name):
+                        if name == "__class__":
+                            return __class__
+                        if name == "self":
+                            return object.__getattribute__(self, "self")
+                        for t in new_class.mro()[1:]:
+                            if t == DelegateSuperclass:
+                                break
+                            if name in t.__dict__:
+                                value = t.__dict__[name]
+                                if get := getattr(value, "__get__", None):
+                                    return get(self.self.this)
+                                return value
+                        return getattr(__graalpython__.super(self.self), name)
+
+                # Wrap all methods so that the `self` inside is always a Java object, and
+                # adapt the globals in the functions to provide a custom super() if
+                # necessary
+                def self_as_java_wrapper(k, value):
+                    if type(value) is not types.FunctionType:
+                        return value
+                    if k in ("__new__", "__class_getitem__"):
+                        return value
+                    if "super" in value.__code__.co_names:
+                        value = types.FunctionType(
+                            value.__code__,
+                            value.__globals__ | {"super": custom_super},
+                            name=value.__name__,
+                            argdefs=value.__defaults__,
+                            closure=value.__closure__,
+                        )
+                    return lambda self, *args, **kwds: value(self.__this__, *args, **kwds)
+                namespace = {k: self_as_java_wrapper(k, v) for k, v in namespace.items()}
+                new_class = type.__new__(mcls, name, bases, namespace)
+                return new_class
+            return type.__new__(mcls, name, bases, namespace)
+
+        def __getattr__(self, name):
+            return getattr(JavaClass, name)
+
+    # A class that defines the required construction for the Java instances, so
+    # the Python code can actually override __new__ to affect the construction
+    # of the Java object
+    class DelegateSuperclass(metaclass=JavaSubclassMeta):
+        def __new__(cls, *args, **kwds):
+            delegate = object.__new__(cls)
+            java_object = polyglot.__new__(JavaClass, *(args + (delegate,)))
+            delegate.__this__ = java_object
+            delegate.__init__(*args, **kwds)
+            return java_object
+
+    return type(name, (DelegateSuperclass,), ns)