+++ scripts/wheelbuilder/darwin/numpy.2.0.2.sh
@@ -38,10 +38,8 @@
 # SOFTWARE.
 
 if [ -n "$GITHUB_RUN_ID" ]; then
-    dnf install -y gcc-toolset-12-gcc-gfortran openblas-devel
-    if [ -n "$1" ]; then
-        scl enable gcc-toolset-12 "pip wheel \"numpy==$1\""
-    else
-        scl enable gcc-toolset-12 "pip wheel numpy"
-    fi
+    brew install gcc openblas
 fi
+
+export CC="gcc-15"
+export CXX="g++-15"